// services/security_monitoring/src/v2/models/SecurityMonitoringThirdPartyRuleCaseCreate.ts
import { AttributeTypeMap } from "@datadog/datadog-api-client";

import { SecurityMonitoringRuleSeverity } from "./SecurityMonitoringRuleSeverity";

/**
 * Case when a signal is generated by a third party rule.
 */
export class SecurityMonitoringThirdPartyRuleCaseCreate {
  /**
   * Name of the case.
   */
  "name"?: string;
  /**
   * Notification targets for each case.
   */
  "notifications"?: Array<string>;
  /**
   * A query to map a third party event to this case.
   */
  "query"?: string;
  /**
   * Severity of the Security Signal.
   */
  "status": SecurityMonitoringRuleSeverity;
  /**
   * A container for additional, undeclared properties.
   * This is a holder for any undeclared properties as specified with
   * the 'additionalProperties' keyword in the OAS document.
   */
  "additionalProperties"?: { [key: string]: any };
  /**
   * @ignore
   */
  "_unparsed"?: boolean;

  /**
   * @ignore
   */
  static readonly attributeTypeMap: AttributeTypeMap = {
    name: {
      baseName: "name",
      type: "string",
    },
    notifications: {
      baseName: "notifications",
      type: "Array<string>",
    },
    query: {
      baseName: "query",
      type: "string",
    },
    status: {
      baseName: "status",
      type: "SecurityMonitoringRuleSeverity",
      required: true,
    },
    additionalProperties: {
      baseName: "additionalProperties",
      type: "any",
    },
  };

  /**
   * @ignore
   */
  static getAttributeTypeMap(): AttributeTypeMap {
    return SecurityMonitoringThirdPartyRuleCaseCreate.attributeTypeMap;
  }

  public constructor() {}
}